+++ packages/tools/utils/src/modules/any-[array-object].js
@@ -5,7 +5,8 @@
 import * as _ from 'lamb';
 
 /**
- * Return a function expecting an array of keys and returning an object with the provided value as value of those keys.
+ * Return a function expecting an array of keys and returning an object with
+ * the provided value as value of those keys.
  *
  * @function
  * @arg {*} value