+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ## `@svizzle/utils` v0.16.0 (next)
 
-- add `arraySumWith`
+- add `arraySumWith`, `updateKeys`
 
 
 ## 20210714